requirements.txt
# Core dependencies for AI PINN project
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
plotly>=5.15.0
pandas>=2.0.0
PyYAML>=6.0
tensorboard>=2.13.0
psutil>=5.9.0